+++ darts/tests/utils/tabularization/test_get_feature_times.py
@@ -1055,7 +1055,7 @@ class TestGetFeatureTimes:
         with pytest.raises(ValueError) as err:
             _get_feature_times(target_series=series, lags=[0], is_training=False)
         assert (
-            "`lags` must be a `Sequence` containing only `int` values less than 0."
+            "`lags` must be a `Sequence` or `Dict` containing only `int` values less than 0."
         ) == str(err.value)
         # `lags_past_covariates` not <= -1:
         with pytest.raises(ValueError) as err:
@@ -1063,7 +1063,7 @@ class TestGetFeatureTimes:
                 past_covariates=series, lags_past_covariates=[0], is_training=False
             )
         assert (
-            "`lags_past_covariates` must be a `Sequence` containing only `int` values less than 0."
+            "`lags_past_covariates` must be a `Sequence` or `Dict` containing only `int` values less than 0."
         ) == str(err.value)
         # `lags_future_covariates` can be positive, negative, and/or zero - no error should be thrown:
         _get_feature_times(